+++ cpp/include/raft/sparse/detail/cusparse_wrappers.h
@@ -16,6 +16,7 @@
 
 #pragma once
 
+#include <cusparse_v2.h>
 #include <raft/error.hpp>
 #include <raft/sparse/detail/cusparse_macros.h>
