+++ cpp/include/raft/sparse/distance/detail/ip_distance.cuh
@@ -20,7 +20,7 @@
 #include <raft/cuda_utils.cuh>
 #include <raft/cudart_utils.h>
 #include <raft/linalg/distance_type.h>
-#include <raft/sparse/cusparse_wrappers.h>
+#include <raft/sparse/detail/cusparse_wrappers.h>
 
 #include <raft/sparse/convert/csr.hpp>
 #include <raft/sparse/convert/dense.hpp>
