+++ cpp/test/sparse/csr_to_dense.cu
@@ -20,7 +20,7 @@
 
 #include <gtest/gtest.h>
 #include <raft/sparse/convert/dense.hpp>
-#include <raft/sparse/cusparse_wrappers.h>
+#include <raft/sparse/detail/cusparse_wrappers.h>
 
 #include <rmm/device_uvector.hpp>
