+++ cpp/test/sparse/csr_transpose.cu
@@ -20,7 +20,7 @@
 
 #include <raft/cudart_utils.h>
 #include <raft/handle.hpp>
-#include <raft/sparse/cusparse_wrappers.h>
+#include <raft/sparse/detail/cusparse_wrappers.h>
 #include <raft/sparse/linalg/transpose.hpp>
 
 #include "../test_utils.h"
