+++ cpp/test/sparse/knn.cu
@@ -19,11 +19,11 @@
 
 #include "../test_utils.h"
 #include <raft/linalg/distance_type.h>
-#include <raft/sparse/cusparse_wrappers.h>
+#include <raft/sparse/detail/cusparse_wrappers.h>
 #include <raft/sparse/selection/knn.hpp>
 
 #include <raft/cudart_utils.h>
-#include <raft/sparse/cusparse_wrappers.h>
+#include <raft/sparse/detail/cusparse_wrappers.h>
 
 namespace raft {
 namespace sparse {
